requirements.txt
colorlog
python-dotenv
openai
flask
flask-cors